Camera Movement Code
Seems like the DriveTrain is working ok with lead and follow. We will
test this out tomorrow when we hook up the motors.

Added in Camera code. Make it so the camera does not move with joystick
unless A button on xbox is held. When A is released the camera stays
where it is.

Signed-off-by: SagesOfRPG <wayne@sagesofrpg.com>